fixed an error that would occur if DSRV was not installed
